+++ CHANGELOG.md
@@ -8,6 +8,10 @@
 - Remove (previously deprecated) `llama_index.token_counter.token_counter` module. See [migration guide](/how_to/callbacks/token_counting_migration.html) for more details on new callback based token counting.
 - Remove `ChatGPTLLMPredictor` and `HuggingFaceLLMPredictor`. See [migration guide](/how_to/customization/llms_migration_guide.html) for more details on replacements.
 - Remove support for setting `cache` via `LLMPredictor` constructor.
+- Update `BaseChatEngine` interface: 
+  - adjust `chat_history` to use `List[ChatMessage]]` as type
+  - expose `chat_history` state as a property
+  - support overriding `chat_history` in `chat` and `achat` endpoints
 
 ## [v0.6.38] - 2023-07-02